requirements.txt
Django==4.0.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
numpy==1.22.3
pandas==1.4.2

